# Optimized sim config filling; avoid reflection.