--- a/godeps.txt
+++ b/godeps.txt
@@ -273,7 +273,6 @@ github.com/ServiceWeaver/weaver/internal/sim
     errors
     fmt
     github.com/ServiceWeaver/weaver
-    github.com/ServiceWeaver/weaver/internal/config
     github.com/ServiceWeaver/weaver/internal/reflection
     github.com/ServiceWeaver/weaver/internal/weaver
     github.com/ServiceWeaver/weaver/runtime