Recall that a simulator has to initialize a set of components. Part of this initialization involves filling in any `weaver.WithConfig`s. Previously, we used the `config.Config` function to do this, but `config.Config` is slow and allocates a lot because it heavily uses reflection. This PR replaces `config.Config` with a more optimized approach that doesn't use any reflection.

```
$ go test -run=$^ -bench=. -count=5 | tee /tmp/baseline.txt
$ go test -run=$^ -bench=. -count=5 | tee /tmp/experiment.txt
$ benchstat /tmp/{baseline,experiment}.txt
name                        old time/op    new time/op    delta
NewWorkload-128               1.30µs ± 1%    1.30µs ± 1%     ~     (p=0.690 n=5+5)
NewSimulator-128              31.6µs ± 2%    26.6µs ± 2%  -15.63%  (p=0.008 n=5+5)
Workloads/NoCallsNoGen-128    40.6µs ± 4%    38.4µs ± 2%   -5.35%  (p=0.008 n=5+5)
Workloads/NoCalls-128         44.8µs ± 1%    40.7µs ± 2%   -9.11%  (p=0.008 n=5+5)
Workloads/OneCall-128         60.0µs ± 2%    56.6µs ± 1%   -5.59%  (p=0.008 n=5+5)

name                        old alloc/op   new alloc/op   delta
NewWorkload-128                 464B ± 0%      464B ± 0%     ~     (all equal)
NewSimulator-128              7.95kB ± 0%    7.61kB ± 0%   -4.33%  (p=0.008 n=5+5)
Workloads/NoCallsNoGen-128    8.53kB ± 0%    8.19kB ± 0%   -4.03%  (p=0.008 n=5+5)
Workloads/NoCalls-128         9.08kB ± 0%    8.70kB ± 0%   -4.18%  (p=0.008 n=5+5)
Workloads/OneCall-128         11.0kB ± 0%    10.6kB ± 0%   -3.44%  (p=0.008 n=5+5)

name                        old allocs/op  new allocs/op  delta
NewWorkload-128                 3.00 ± 0%      3.00 ± 0%     ~     (all equal)
NewSimulator-128                70.0 ± 0%      53.0 ± 0%  -24.29%  (p=0.008 n=5+5)
Workloads/NoCallsNoGen-128      85.0 ± 0%      68.0 ± 0%  -20.00%  (p=0.008 n=5+5)
Workloads/NoCalls-128           99.0 ± 0%      82.0 ± 0%  -17.17%  (p=0.008 n=5+5)
Workloads/OneCall-128            138 ± 0%       121 ± 0%  -12.32%  (p=0.008 n=5+5)
```